21-09-2025 :Refactorizacion de componentes ActualClima
