--- styles/TempeActual.css
@@ -1,372 +1,0 @@
-/* ==============================
-   Contenedor principal
-   ============================== */
-.weather-container {
-  max-width: 420px;
-  margin: 30px auto;
-  padding: 24px;
-  border-radius: 16px;
-  background: #f9fafb; /* Fondo neutro */
-  color: #333;
-  font-family: "Segoe UI", Roboto, sans-serif;
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.08);
-  text-align: center;
-  animation: fadeIn 0.5s ease-in-out;
-}
-
-/* Animación de entrada */
-@keyframes fadeIn {
-  from { opacity: 0; transform: translateY(10px); }
-  to { opacity: 1; transform: translateY(0); }
-}
-
-.weather-container h1 {
-  font-size: 1.5rem;
-  margin-bottom: 18px;
-  font-weight: 600;
-  color: #222;
-}
-
-/* ==============================
-   Bloques de información
-   ============================== */
-.weather-main,
-.weather-description {
-  background: #ffffff;
-  border-radius: 12px;
-  padding: 14px 16px;
-  margin-bottom: 12px;
-  border: 1px solid #e5e7eb;
-  text-align: left;
-}
-
-.weather-main p,
-.weather-description p {
-  margin: 6px 0;
-  font-size: 0.95rem;
-  color: #444;
-}
-
-/* ==============================
-   Input de ciudad
-   ============================== */
-.city-input input {
-  width: 100%;
-  padding: 10px 14px;
-  border: 1px solid #d1d5db;
-  border-radius: 10px;
-  font-size: 0.95rem;
-  outline: none;
-  background: #fff;
-  color: #333;
-  transition: border 0.2s, box-shadow 0.2s;
-}
-
-.city-input input::placeholder {
-  color: #9ca3af;
-}
-
-.city-input input:focus {
-  border-color: #3b82f6;
-  box-shadow: 0 0 0 2px rgba(59, 130, 246, 0.25);
-}
-
-.city-input small {
-  display: block;
-  margin-top: 8px;
-  font-size: 0.8rem;
-  color: #6b7280;
-}
-
-/* ==============================
-   Botón (para errores o recarga)
-   ============================== */
-.weather-container button {
-  margin-top: 14px;
-  padding: 10px 18px;
-  border: none;
-  border-radius: 8px;
-  background: #3b82f6;
-  color: #fff;
-  font-size: 0.9rem;
-  font-weight: 500;
-  cursor: pointer;
-  transition: background 0.2s, transform 0.2s;
-}
-
-.weather-container button:hover {
-  background: #2563eb;
-  transform: translateY(-1px);
-}
-
-.weather-container button:active {
-  transform: translateY(0);
-}
-
-
-
-/* Modal  */
-/* Agregar estos estilos a tu archivo TempeActual.css */
-
-.weather-container {
-  max-width: 800px;
-  margin: 0 auto;
-  padding: 20px;
-  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-}
-
-/* Header Section */
-.header-section {
-  margin-bottom: 30px;
-}
-
-.header-section h1 {
-  display: flex;
-  align-items: center;
-  gap: 15px;
-  margin-bottom: 20px;
-  color: #2c3e50;
-}
-
-.country-badge {
-  background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-  color: white;
-  padding: 4px 12px;
-  border-radius: 20px;
-  font-size: 0.7em;
-  font-weight: 500;
-  text-transform: uppercase;
-  letter-spacing: 0.5px;
-}
-
-/* Search Container */
-.search-container {
-  position: relative;
-  width: 100%;
-  max-width: 400px;
-}
-
-.input-wrapper {
-  display: flex;
-  border-radius: 25px;
-  overflow: hidden;
-  box-shadow: 0 4px 15px rgba(0, 0, 0, 0.1);
-  background: white;
-}
-
-.city-search-input {
-  flex: 1;
-  padding: 12px 20px;
-  border: none;
-  outline: none;
-  font-size: 16px;
-  background: transparent;
-}
-
-.city-search-input::placeholder {
-  color: #bdc3c7;
-}
-
-.search-button {
-  padding: 12px 16px;
-  border: none;
-  background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-  color: white;
-  cursor: pointer;
-  font-size: 16px;
-  transition: all 0.3s ease;
-}
-
-.search-button:hover {
-  transform: scale(1.05);
-}
-
-/* Modal de Sugerencias */
-.suggestions-modal {
-  position: absolute;
-  top: 100%;
-  left: 0;
-  right: 0;
-  background: white;
-  border-radius: 10px;
-  box-shadow: 0 8px 25px rgba(0, 0, 0, 0.15);
-  z-index: 1000;
-  max-height: 300px;
-  overflow-y: auto;
-  border: 1px solid #e1e8ed;
-  margin-top: 5px;
-}
-
-.suggestion-item {
-  padding: 12px 20px;
-  cursor: pointer;
-  border-bottom: 1px solid #f8f9fa;
-  transition: all 0.2s ease;
-}
-
-.suggestion-item:hover {
-  background-color: #f8f9fa;
-  transform: translateX(5px);
-}
-
-.suggestion-item:last-child {
-  border-bottom: none;
-}
-
-.suggestion-item.loading,
-.suggestion-item.no-results {
-  color: #6c757d;
-  cursor: default;
-  text-align: center;
-  font-style: italic;
-}
-
-.suggestion-item.loading:hover,
-.suggestion-item.no-results:hover {
-  background-color: transparent;
-  transform: none;
-}
-
-.suggestion-main {
-  font-size: 16px;
-  color: #2c3e50;
-  margin-bottom: 2px;
-}
-
-.suggestion-country {
-  font-size: 12px;
-  color: #7f8c8d;
-  text-transform: uppercase;
-  letter-spacing: 0.5px;
-}
-
-/* Weather Main Section */
-.weather-main {
-  display: grid;
-  grid-template-columns: auto 1fr;
-  gap: 30px;
-  align-items: center;
-  background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-  color: white;
-  padding: 30px;
-  border-radius: 20px;
-  margin-bottom: 25px;
-  box-shadow: 0 10px 30px rgba(102, 126, 234, 0.3);
-}
-
-.temperature-section {
-  text-align: center;
-}
-
-.main-temp {
-  font-size: 4rem;
-  font-weight: 300;
-  line-height: 1;
-  margin-bottom: 5px;
-}
-
-.feels-like {
-  font-size: 0.9rem;
-  opacity: 0.9;
-}
-
-.weather-details {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(120px, 1fr));
-  gap: 15px;
-}
-
-.detail-item {
-  text-align: center;
-  padding: 15px;
-  background: rgba(255, 255, 255, 0.1);
-  border-radius: 10px;
-  backdrop-filter: blur(10px);
-}
-
-.detail-item .label {
-  display: block;
-  font-size: 0.8rem;
-  opacity: 0.8;
-  margin-bottom: 5px;
-}
-
-.detail-item .value {
-  display: block;
-  font-size: 1.2rem;
-  font-weight: 600;
-}
-
-/* Weather Description */
-.weather-description {
-  background: white;
-  border-radius: 15px;
-  padding: 20px;
-  margin-bottom: 25px;
-  box-shadow: 0 5px 15px rgba(0, 0, 0, 0.1);
-}
-
-.condition {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 15px;
-}
-
-.condition-text {
-  font-size: 1.2rem;
-  color: #2c3e50;
-  font-weight: 500;
-}
-
-.weather-icon {
-  width: 50px;
-  height: 50px;
-}
-
-/* Loading Spinner */
-.loading-spinner {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 200px;
-  font-size: 1.2rem;
-  color: #667eea;
-}
-
-/* Responsive Design */
-@media (max-width: 768px) {
-  .weather-container {
-    padding: 15px;
-  }
-  
-  .weather-main {
-    grid-template-columns: 1fr;
-    text-align: center;
-    gap: 20px;
-  }
-  
-  .main-temp {
-    font-size: 3rem;
-  }
-  
-  .weather-details {
-    grid-template-columns: repeat(2, 1fr);
-  }
-  
-  .header-section h1 {
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 10px;
-  }
-}
-
-@media (max-width: 480px) {
-  .weather-details {
-    grid-template-columns: 1fr;
-  }
-  
-  .main-temp {
-    font-size: 2.5rem;
-  }
-}
